requirements.txt
aiobotocore
aiosu
asyncpg
blend_modes
databases
discord
html2image
httpx
numpy
Pillow
python-dotenv
selenium
slider
